refactor: rename navigation to header
