+++ src/components/Header/header.test.js
@@ -1,8 +1,8 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
-import Navigation from './Navigation'
+import Header from './Header'
 
 it('renders Navigation component', () => {
   const div = document.createElement('div')
-  ReactDOM.render(<Navigation />, div)
+  ReactDOM.render(<Header />, div)
 })
